+++ Planner/app/src/main/java/rdatu/android/cyscorpions/com/projectplanner/view/activities_fragments/ScheduleTaskActivity.java
@@ -14,7 +14,7 @@ import android.widget.LinearLayout;
 import android.widget.Toast;
 
 import rdatu.android.cyscorpions.com.projectplanner.R;
-import rdatu.android.cyscorpions.com.projectplanner.controller.modelmanagers.TaskManager;
+import rdatu.android.cyscorpions.com.projectplanner.controller.model_managers.TaskManager;
 import rdatu.android.cyscorpions.com.projectplanner.model.objects.Tasks;
 import rdatu.android.cyscorpions.com.projectplanner.view.dialogs.DatePickerDialog;
 import rdatu.android.cyscorpions.com.projectplanner.view.dialogs.OverwriteDialog;
